FastAPI/learning/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
alembic==1.14.0
